[CI] Add ARM wheels for macOS
CIBuildWheels allows cross compilation to ARM on mac platforms, very
clever indeed.
To enable this you need to specify the archs explicitly, which this PR
does. This should allow our apple silicon users to not have to go
fiddling with override arch settings when they want to use OpenAssetIO.

The "normal" way of doing this would be as we had it setup (without the
exception of course), but on a arm machine, but github only supports
self hosted runners for those. (At least for now, silicon runners are
expected later this year).

We could also consider doing this for ubuntu, as cibuildwheels lets you
do a build under emulation (using QEMU), but I fear this will be
extremely slow. Worth considering though.

Signed-off-by: Elliot Morris <elliot.morris@foundry.com>