[Core] Add fmtlib dependency
Closes #1070. fmtlib allows for easy, safe string substitution in C++,
adding advanced formatting functionality (e.g. rounding), and reducing
bloat.

So add as a new private build-only header-only dependency.

Using header-only increases compile times, but subjectively it's not
noticeable, and has the benefit of avoiding a possible public library
dependency (if OpenAssetIO is built as a static library).

Tweak a single existing, tested, log string to make use of libfmt, to
ensure basic usage is validated.

Signed-off-by: David Feltell <david.feltell@foundry.com>