[Core] Remove backdoor access to Manager internals
These were added when we had to re-implement ManagerFactory in Python, so we didn't lose access to
methods not defined in the core C++ API.  We now have the full surface area in C++. Removing them
discourages shenanigans.

We didn't mind losing the test coverage of the `Manager` constructor, as it is implicitly checked
in every case where we use a mock interface, and modify its responses as part of a test case.

Closes #26

Signed-off-by: Tom Cowland <tom@foundry.com>