[Core] Migrate to new exception hierarchy
Part of #1071

It is useful for hosts to be able to branch on specific types of
exception in order to customize their mitigation. In particular, it is
useful for hosts to know if an exception comes from the OpenAssetIO
ecosystem.

Migrate non-batch python-implemented exceptions to be C++ based.
In addition, exception hierarchy has been redesigned to better reflect
current understanding of error handling needs.

Notably, all exceptions now share a base type, `OpenAssetIOException`,
allowing hosts the ability to catch on a single expectable exception
type around their implementation.

As part of this work, several library exception types that were thrown
in the core api have been migrated to the new appropriate OpenAssetIO
exception types. In addition, some expected thirdparty exceptions have
been caught and rethrown as OpenAssetIO exceptions.

BatchElementExceptions have not been migrated as part of this work and
will be done in a subsequent commit. However, BatchElementError and
Exceptions have been moved under an errors directory in the source tree

Signed-off-by: Elliot Morris <elliot.morris@foundry.com>