[Core] Improve `defaultManagerForInterface` error reporting
Attempting to load a default config from a dir throws (for good
reasons). The previous error was from deep in the bowels of TOML++, and
didn't make it particularly obvious what was wrong. As
OPENASSETIO_PLUGIN_PATH should be set to directories, there is a
relatively high chance they could get mixed up. This hopefully clarifies
matters some.

Signed-off-by: Tom Cowland <tom@foundry.com>