[Core] Expanded batch exception messages
Part of #1073

As part of an exception refresh, move batch errors and exceptions into
the errors namespace. Also rearrange the source tree to reflect this
with an errors directory.

In order to provide easier debuggability, automatically
construct an error message with helpful information in the case
where the user calls an exceptional convenience.

Over the course of implementation, it became clear that the
`BatchElementException` hierarchy introduced more burden than benefit,
so all the subtypes have been removed, favouring a single simple
exception type that is always thrown when a `BatchElementError` is
emitted in exceptional convenience cases.

Signed-off-by: Elliot Morris <elliot.morris@foundry.com>