[Core] Add hasCapability method
Closes #1119.

Adds the `hasCapability` method to the `Manager`/`ManagerInterface` in
order to provide hosts with a way of introspecting on which methods are
implemented.

This comes from a slight shift in our thinking around default methods,
where we are no longer going to return "empty" values to mean
unimplemented (as empty values are a legitimate return state) but
will instead throw to inform hosts that methods are unimplemented.

This necessarily requires an introspection mechanism to avoid a
try-and-see approach.

In order to have middleware-specific capabilities, split the capability
`enum class`es between `Manager` and `ManagerInterface`, where `Manager`
has a subset of those in `ManagerInterface`. Hence place the root
`Capability` enum under the `internal` namespace, so that the values
can be shared - useful for e.g. indexing into a `kCapabilityNames` array
for auditing/logging/UI.

Signed-off-by: Elliot Morris <elliot.morris@foundry.com>
Signed-off-by: David Feltell <david.feltell@foundry.com>