[CI] Fix Pylint scanning of `tests` directory
Closes #807. The `--recursive` option added in Pylint v2.13.5 allows
Pylint to use a directory of test modules that are not structured as a
Python package, i.e. it avoids errors like

> No such file or directory: [...] __init__.py

This particular error, frustratingly, does not actually fail the check
on CI, so we built up quite a backlog of Pylint warnings in the `tests`
directory.

So pin to the latest v2 of Pylint, which supports this feature, and use
it in the CI config.

Signed-off-by: David Feltell <david.feltell@foundry.com>