[Core] Remove non-paged realtionship query methods
Having 4 different methods to implement was somewhat onerous for the
manager. Especially as they had to ensure consistent results between all
variations.

In practice, we believe the overhead of the pager is not significant
enough to warrant a first-class path in the `ManagerInterface` at this
point. We can always add this in at a later date should such an
optimization point prove necessary.

This brings these methods in line with other (currently) Manager-only
conveniences.

Closes #1125

Signed-off-by: Tom Cowland <tom@foundry.com>