[Core] Move `Access` root enum to new `internal` namespace
Part of #1119. In preparation for adding new enumerations that share
values between them, establish a pattern of placing such shared values
in a private `internal` namespace, starting with the `Access` enum.

Use `std::size_t` as the underlying type of the enum (and propagate to
the `enum class`es derived from it) so that `Access` values can be used
to index into the `kAccessNames` array without need to cast.

Signed-off-by: David Feltell <david.feltell@foundry.com>