[Core] Finalize ManagerInterface abstract methods
Part of #163

Makes most methods throw `NotImplementedException`, when called without
a manager implementation. This is primarily to support the future
implementation of a hybrid plugin system, which makes language specific
pure-virtual/abstract mechanisms impossible.

Adds hopefully helpful error messages to these exceptions, instructing
in the correct usage of `hasCapability` in the case that the host
encounters these exceptions.

API compliance suite check for `updateTerminology` always runs because
it does not depend on a fixture, so use `hasCapability` in that one
case. Future work will use `hasCapability` in the compliance suite more
broadly.

Signed-off-by: Elliot Morris <elliot.morris@foundry.com>
Signed-off-by: David Feltell <david.feltell@foundry.com>