[Python] Release the GIL in non-trivial bound methods
Closes #554. For C++ function bodies that may be long running, releasing
the GIL is good practice so that other Python threads can continue
whilst the C++ function runs.  For C++ functions that ultimately call
back out to Python in a separate thread, it is _crucial_ that the GIL is
released in order to prevent deadlock.

So release the GIL during the function body of all such bound methods.

As well as the middleware wrapper classes (`Manager` et al.), the
`*Interface` class bindings (`ManagerInterface` et al.) also release the
GIL, since these may be called directly in legitimate use-cases (e.g. a
GRPC wrapper).

Most of the work here is tests. Tests are added that will (hopefully)
catch any newly added bound methods and force us to consider the GIL.

Signed-off-by: David Feltell <david.feltell@foundry.com>