[CI] More robust openassetio uninstall regex guard
Fix regex used to detect if TraitGen's `requirements.txt` tries to
uninstall `openassetio`.

In OpenAssetIO/OpenAssetIO-TraitGen#51 we had to remove the PyPI
dependency on `openassetio` temporarily, since the project now depends
on the breaking changes in the latest OpenAssetIO main branch. So
`openassetio` was commented out in `tests/requirements.txt` in that
project.

We now have two valid cases, either TraitGen's `requirements.txt` tries
to install `openassetio` and short-circuits (saying it's already
installed), or it doesn't try at all.

So update the guard regex to catch either case, i.e. assert that `pip`
does not attempt to uninstall `openassetio`, for whatever reason.

Signed-off-by: David Feltell <david.feltell@foundry.com>