[CI] Add TraitGen C++ integration tests
Closes #841. Since OpenAssetIO/OpenAssetIO-TraitGen#11 (also see
follow-up OpenAssetIO/OpenAssetIO-TraitGen#17) there are C++ specific
tests for TraitGen.

We were running the TraitGen tests alongside other Python-only tests
that use OpenAssetIO installed as a Python pip package, which the
TraitGen CMake CTest project cannot use to build against.

So add a separate test run of TraitGen under the "CMake package"
integration tests. Only run the tests labelled `ctest`, since the tests
for Python codegen are still executed under the "Python e2e" tests.

Since we use `awk` to assert that TraitGen test dependencies don't
overwrite OpenAssetIO, we must use `bash` as the shell for all
platforms. This means `vcvarsall.bat` can't be used on Windows. So
switch to using the default generator for the platform (rather than
Ninja), which on Windows will be MSVC and so not requiring
`vcvarsall.bat`.

Signed-off-by: David Feltell <david.feltell@foundry.com>