[CI] Add Build Archive CI script.
Add archive behaviour to merges to main, so OpenAssetIO is built on all
of our runner platforms and archived in github.

Intent is to unlock much better CI behaviour for us, as we are spending
a lot of time and effort building OpenAssetIO across dependent repos,
as well as just being generally convenient ot have.

Signed-off-by: Elliot Morris <elliot.morris@foundry.com>