[Python] Decorate override macros with exception converter
Part of #947. Augment pybind11 override macros such that
OpenAssetIO-specific exceptions, thrown in the Python override of a C++
virtual method, will be converted from a Python exception (captured by
pybind11 as  an `error_already_set` exception) to an exception that can
propagate and be caught either in C++, as the corresponding C++
exception type, or back in Python.

Move macros under their own header, since there's a fair chunk of code
now, and since not every Python binding needs them.

Move Python->C++ exception translation utils to a shared private header.
Primarily this was so that the functions can be used in the
openassetio-python-bridge target. However, on inspection there isn't
actually a current use-case for this in the Python bridge. This change
still seems worthwhile for future-proofing, and since it's a little
neater.

Add a section to our coding standards notifying contributors that they
should always use the `OPENASSETIO_`-prefixed pybind11 override macros.

Signed-off-by: David Feltell <david.feltell@foundry.com>