[Core] Context default-construct non-null locale
Part of #1153. A `Context` with a null `locale` could cause an error
(exception/segfault) when copy-constructed, e.g. through
`createChildContext`. In theory a `Context` should never be constructed
directly, but it is possible (and used a lot in tests, for example).

So if no `locale` argument is given to the `Context` constructor in C++,
default to constructing an empty, rather than null, `TraitsData`.

For Python we cannot use a non-null default constructor argument,
because the same instance is re-used whenever a `Context` is created,
i.e. changes to the default-constructed `locale` would be reflected in
the next brand new `Context`.

Instead, remove the default value for the `locale` and disallow `None`,
and provide an alternative no-argument constructor to handle the default
construction case.

Signed-off-by: David Feltell <david.feltell@foundry.com>