[Python] Add value to `EntityReference` `repr` output
We opted not to go for `EntityReference("theStr")` as it gives the
impression they may be directly contructed. Outside of tests, this is
discouraged, as `manager.createEntityReference` shoud be used instead.

The present implementation avoids quoting the reference and skirts the
issue of escaping angle brackes, as this should only be used for debug
introspection.

Signed-off-by: Tom Cowland <tom@foundry.com>