[CI] Enforce usage of OPENASSETIO pybind11 macros
Part of #947. In order to translate exceptions from Python->C++ (as well
as work around other pybind11 issues), we must always use the
`OPENASSETIO_`-prefixed `PYBIND11_OVERRIDE` macros, which augment
pybind11's implementation with our patches.

So add a simple CI `grep` check to enforce this.

Re-use the existing check for "OAIO" and re-purpose as a generic
"disallowed phrases" check. This may also be useful for #1112.

Signed-off-by: David Feltell <david.feltell@foundry.com>