[Build] Add cmake-lint check to linter builds
Closes #1132. Ensure a consistent style in CMake files. This is
especially important as we start to attract outside collaborators.

Following the pattern of other linters, e.g. `cpplint`, add an optional
build target that executes `cmake-lint`, and wire that target as a
dependency of all targets configured using the
`openassetio_set_default_target_properties` function.

Add to `linter-requirements.txt` and enable the cmake-lint check in
linting CMake presets, so that it is installed and runs on CI (and
well-behaved local dev machines).

Signed-off-by: David Feltell <david.feltell@foundry.com>