[Core] Add entityTraits
Closes #31.

In order to support generic import style workflows (eg: retrieving an
entity reference from a drop event), it is necessary for a host to
determine the characteristics of an entity.

This method allows the appropriate reader/loader/etc. to be created. At
present, it is not possible to determine an entities traits from an
entity ref.

When publishing it also allows introspection of the expected traits so
behaviour can be adapted to ensure successful operation.

Unlike managementPolicy the resulting trait set should contain all known
traits that represent qualities of the entity in question (or future
entity), it should not omit any traits that have no properties, or the
manager is unable to provide data for.

Passing the result of this through managmentPolicy would tell you which
traits a manager may be able to populate/persist.

So add new core API method `entityTraits` including associated
capability `kEntityTraitIntrospections` and access enum
`EntityTraitsAccess`.

The `kEntityTraitIntrospections` capability is a required capability and
so is not exposed to hosts (i.e. via `Manager`), and is asserted during
manager plugin initialization.

API compliance suite updated to check error cases and batch processing.

Examples.dox contrived example added demonstrating possible usage for a
drag-and-drop system. Various other docstrings updated to add
cross-references to `entityTraits`.

Signed-off-by: David Feltell <david.feltell@foundry.com>