[Python] Fix HybridException base class RTTI
Closes #1225. The `HybridException` mechanism used multiple inheritance
to allow an exception to be caught as either an OpenAssetIO exception or
a pybind11 exception, allowing it to propagate from Python through C++
and back out to Python without losing information.

However, this entailed diamond inheritance, where the two branches of
the inheritance hierarchy had a root base class of `std::exception`.

This meant that a `HybridException` could not be caught as a
`std::exception` due to the ambiguity. Frustratingly, this manifests as
if `std::exception` wasn't part of the hierarchy at all, at least on
GCC, and the exception would simply fall through the `catch`.

So rework the `HybridException` mechanism to use composition, rather
than inheritance. I.e. we compose the Python exception proxy object
`pybind11::error_already_set` rather than add it to the base class list.

This means we need additional explicit detection of `HybridException`s
when translating exceptions from C++ to Python, to ensure we re-raise
the original Python exception, rather than creating a new Python
exception and losing e.g. stack trace data.

Luckily, when we `throw` a `pybind11::error_already_set` from within a
pybind11 `register_exception_translator` callback, it's handled by
pybind11 as we expect and the original Python exception is restored.

On MacOS (clang / libc++), RTTI is more strict than other platforms
(pointer vs. string based) and requires exception symbols to be visible
in order to be thrown in `_openassetio_test` and caught (via pybind11
global exception translator callback) in `_openassetio`.

So export the `HybridException` template class symbols.

Use the handy `PYBIND11_EXPORT` macro, which is the pybind11 documented
way to share symbols across extension modules. This appears to be a
trimmed down version of regular export macros - e.g. it has no
`dllimport` branch for Windows, avoiding the need for a `.lib`.

Ad-hoc experimentation revealed that this symbol export should not be
necessary - it's only required because of the `_openassetio_test` tests.
Future work to refactor these tests will allow us to remove this
unnecessary symbol export.

Signed-off-by: David Feltell <david.feltell@foundry.com>