[CI] Publish release artifacts.
Publish build artifacts as assets on release page #1201

On release, build an extended matrix of C++ builds, and push them
to the release page.

This is good for downstream users, but will also allow us to pin to
releases in our own CI easier, in the case we need C++ versions.

Signed-off-by: Elliot Morris <elliot.morris@foundry.com>