[Core] Remove unnecessary export of HybridException
Part of #1237.

We needed to export the `HybridException` symbol in order for the
`_openassetio_test` test-specific Python extension module to work under
Clang (i.e. MacOS).

Now that the `_openassetio_test` module has been removed in favour of an
alternative approach to test-specific bindings, we can make this symbol
private, as it should be.

Signed-off-by: David Feltell <david.feltell@foundry.com>