[Tests] Add swift-url path<->URL conversion test cases
Part of #1246.

There are many many corner cases when converting paths to/from URLs
(especially on Windows). Thankfully the swift-url project includes a
JSON database containing 1300+ test cases. We intend to implement our
functionality such that our initial implementation satisfies these test
cases.

The only modification of the JSON from upstream is the addition of
license and copyright information and attribution to the original
source, found at the top of the JSON file.

In order to test each entry of the swift-url JSON path<->URL test
database we ideally need a way to treat each case as an independent test
unit.

Unfortunately, pytest has no built-in functionality for parametrizing
test cases based on data loaded from disk.

So bring in the `pytest-subtests` plugin for pytest, which allows a
single test case to be treated like several test cases in a similar way
to `unittest`'s `self.subTest(...)`.

Signed-off-by: David Feltell <david.feltell@foundry.com>