[Core] Add FileUrlPathConverter utility class
Closes #1246. Traits generally use URLs not paths (as we're dearly
trying to get away from the filesystem being everything).

There are few implementations of file path<->URL conversion, and the
ones that exist are either not available in C/C++, are bundled with a
huge dependency (e.g. Qt) or have different behaviour depending on the
platform.

Condensed acceptance criteria:
* Performant, since such transformations are likely to appear on hot
  code paths.
* Support for converting Windows paths on POSIX hosts, and vice versa,
  since servers may well be running on a different plaform than they are
  generating data for.
* Support for UNC Windows paths - required for addressing network shares
  and used to overcome restrictions on other paths (e.g. `\\?\`-prefixed
  paths are not subject to MAX_PATH length limitations).
* Compliant `file://` URLs to WHATWG standard - the modern standard
  adhered to by most browsers.
* Does not re-invent the wheel where there is a de-facto standard
  implementation available.
* Has Python bindings.
* Uses UTF-8 throughout.

Unfortunately we couldn't find any platform-agnostic Windows and POSIX
path parsing libraries.

However, we can use the Ada library for the URL parsing, construction,
(some) validation, and percent encoding.

To simplify the path parsing and normalisation code, we make use of the
PCRE2 regular expression library, so that string searches and
substitutions can be written in concisely, yet executed with high
performance.

However, this brings with it a cost in the form of regular expression
compilation. This cost is front-loaded into the constructor of the
`FileUrlPathConverter` class, so that hot code paths can construct a
single instance, then re-use the pre-compiled regexes for multiple
path<->URL conversions.

There are extensive test cases (=1332) thanks to a JSON database from
the swift-url project. The presented implementation aims to match those
test cases without modification. As such, some functionality that could
be supported is not, and some compromises in code structure are made (in
particular, the priority of error cases when an input has more than one
error).

Unsupported functionality is detailed in the public docstrings, and may
be revisited in future.

Since the pre-compiled regexes are stateful, they are bundled into
multiple `struct`s alongside functions that use them. The division into
`struct`s is dictated by dependency injection. I.e. common functionality
is bundled together and propagated via constructor arguments to more
specific components, eventually arriving at the handler class for a
particular type of path/URL.

Signed-off-by: David Feltell <david.feltell@foundry.com>