[Doc] Add DR23, Traits and Specs versioning
Add the decision record for traits and specifications versioning.
This is a vast topic, this document is an attempt
to provide a tour through our thoughts, whilst
making a large amount of assumptions so we can
try to talk only about one or two topics. For that
reason, it's a little different from a standard decision
record.

Signed-off-by: Elliot Morris <elliot.morris@foundry.com>