[Core] Add kRequired and kManagerDriven access modes
Closes #1209.

Managers may require the data for certain properties to be present in
order to publish an entity. The core API also defines the `preflight` ->
`resolve` -> `register` workflow for when a Host wishes or requires to
derive some of an entity's properties from the Manager during publish.
For example - the location to store a render, or what frame range to
export.

There was however, no way for a Host to query ahead of time which traits
the Manager requires or may be able to resolve for future entities. This
means that the Host has no way to present a coherent user experience
without "trying first".

Using `managementPolicy` to provide this data has the benefit that as
the method does not take an Entity Reference, it is a broad
application-level query which allows Hosts to preemptively manage user
experience if they are not able to provide data a manager needs or the
manager isn't able to provide data the host needs. The resulting low
call count hopefully offsets that several calls may be needed to define
final behaviour.

It also constrains this business logic to specific entity trait sets,
which simplifies implementation as it removes the need (or possibility)
for this to vary between entities of the same trait set, which feels
beneficial in what is an already complex working environment.

It is left to Trait definitions to define which particular properties
are required or optional. A trait being indicated as
supported/required/derivable by `managementPolicy` would, by contract,
require that Hosts provide those properties, and/or Managers would
populate them during `resolve`.

So add the `kRequired` access mode for querying the subset of traits
whose required properties must be set in order for a publish to
succeed. Add the `kManagerDriven` access mode for querying the subset
of traits the manager is capable of providing to the host for use when
creating and publishing new content.

Hence rename `ResolveAccess.kWrite` to `kManagerDriven`, for consistency
with the corresponding `managementPolicy` query.

Adapt example docs to make use of this newly added workflow.

Make minimal changes to API compliance suite, pending #1240.

Signed-off-by: David Feltell <david.feltell@foundry.com>