[Tests] Refactor common batch method tests
Part of #1239.

The test code for conveniences had ballooned massively because it's a
large combinatorial matrix. This makes it hard to add additional
methods (because you have to write lots of duplicates).

So add a `BatchFirstMethodTest` base class with common assertions to be
used for testing the various batch methods. Luckily their call
signatures are all fairly similar.

Signed-off-by: David Feltell <david.feltell@foundry.com>