[Core] Move Manager convenience overloads
Part of #1239.

The member functions of the `Manager` class include several overloads
of the primary batch-first callback-based member functions, in order to
provide a convenient terser interface for users that do not need the
full flexibility afforded by the primary signatures.

The definition of these overloads bloat the source file, making it
harder to read the core functionality at a glance. There are also many
more convenience overloads to come, which will bloat the source file
even more.

So move these convenience overloads to a separate source file.

Signed-off-by: David Feltell <david.feltell@foundry.com>