[Core] pathFromUrl Windows MAX_PATH limit handling
Closes #1257. A common problem in real-world scenarios is the
inconsistent handling of long file paths on Windows hosts/applications.

`pathFromUrl` would convert a Windows file URL to either a Windows drive
path or a standard UNC share path, irrespective of the length of the
path.

In addition, `pathToUrl` loses any information about the type of (UNC)
path provided to it.

This means that e.g. a path longer than the Windows `MAX_PATH` limit fed
into `pathToUrl` as a `\\?\...` path will come back out via
`pathFromUrl` as a standard `C:\...` or ``\\host\share\...` path, which
might be rejected by the Windows API, depending on the particular
Windows machine configuration and application where it is used.

So, if the Windows path parsed from a URL exceeds the `MAX_PATH` limit
(259 chars + null terminator), prefix it so that it becomes an
equivalent UNC non-normalised device path.

Signed-off-by: David Feltell <david.feltell@foundry.com>