[Docs] Add DR for `managementPolicy` design
Adds a (retrospective) DR that attempts to capture the rationale behind
the current design of `managementPolicy`, and why it is constrained by
trait set/access. Much of this information has been lost over the years,
so this aims to help clarify prior decision's and inform any future
iteration so the tradeoffs are better understood.

Signed-off-by: Tom Cowland <info@tomcowland.com>