[Core] Movable ManagerImplementationFactoryInterface
Part of #1115.

A protected member is against C++ Core Guidelines C.133 because it
exposes invariants. We reasoned that because it's `const` and an
abstract interface, then the argument doesn't hold (and C.134 seems to
back this up).

However, a `const` member means the type (and any subclass) is no longer
move-constructible. Being able to move construct is handy, if only
because it provides a workaround to the private constructor being
inaccessible to `std::make_shared` in our various `make` factory
functions.

So for `ManagerImplementationFactoryInterface::logger_`, remove the
`const`, make it private, and expose as a `protected` member function
instead. Bind this as a `property` in Python, so no existing code needs
to change.

Signed-off-by: David Feltell <david.feltell@foundry.com>