[Core] Add CppPluginSystemManagerImplementationFactory
Part of #1115. Add the `CppPluginSystemManagerImplementationFactory`
class and can hence be used by hosts to load C++ plugins via a
`ManagerFactory`.

Implementation largely mirrors the implementation of the
`PythonPluginSystemManagerImplementationFactory`. This is perhaps not
the neatest implementation, but consistency has been prioritised. Any
future improvements should be applied to both C++ and Python
implementations.

Signed-off-by: David Feltell <david.feltell@foundry.com>