[Core] Add LoggerInterface.isSeverityLogged
Closes #1014. Constructing log strings can be expensive, especially in
hot code paths. Doing so only for the log severity level to be such that
the string is never printed is suboptimal, to say the least.

So add a new method `isSeverityLogged`, which returns `true` by default,
but can be overridden by subclasses in order to hint whether a log
string provided to the logger is likely to be output, and so allow users
to short-circuit complex string construction logic.

Update `SeverityFilter` to implement the new method, and make use of the
method of the wrapped logger in determining whether a message should be
filtered.

Signed-off-by: David Feltell <david.feltell@foundry.com>