[Docs] Consider dependent projects in release process
This came out of a team retrospective action to address an issue whereby
OpenAssetIO core was released with a breaking change, but the
corresponding update to BAL (Basic Asset Library) was not yet ready.
This meant there was a non-trivial period where the latest published
version of BAL was incompatible with the latest published version of
OpenAssetIO core. This then had a knock-on effect to host integrators,
who could not use BAL to test conformance with the latest OpenAssetIO
core API.

So update our release process so that we always consider if other
projects in the ecosystem must be released simultaneously, and if so
ensure they are ready for release, before finalising the OpenAssetIO
core release.

Signed-off-by: David Feltell <david.feltell@foundry.com>