[Tests] Don't outout IPO check failure log
Part of #1085. On some systems interprocedural (link time) optimisation
is not supported - e.g. in the openassetio-build Docker image. When
support is not available a long scary looking failure log is output. If
there is a failure for another reason (i.e. a test failure) then this
log obscures the true error.

Signed-off-by: David Feltell <david.feltell@foundry.com>