[CI] Add abidiff check
Part of #1085. Add a CI job that runs the new ABI CTest targets in the
VFX reference platform `openassetio-build` Docker image.

If this job fails, it means the ABI of one of the C/C++ libraries has
changed, and we need to decide whether to revert the changes, accept the
change as harmless, or accept the break and update the release notes.

Signed-off-by: David Feltell <david.feltell@foundry.com>