[Lint] Disable Clang-Tidy chained-comparison check in tests
Discovered during #1169. As of Clang-Tidy 18, a
bugprone-chained-comparison check was added that flags e.g. `a <= b ==
c`, which, due to operator overloading, Catch2 `CHECK` macros use.

Note that at time of writing CI is unaffected, since we're stuck on
Clang-Tidy 12. So this is currently just to satisfy local (IDE)
diagnostics.

Signed-off-by: David Feltell <david.feltell@foundry.com>