[Core] Add `substitute` utility function
Closes #1213. Many strings communicated between a host and a manager may
contain placeholder tokens that are intended to be substituted by the
host before use, e.g. an animated image sequence on disk as several
files similarly named, except for a frame number, could be referred to
collectively using `image.{frame}.exr`.

However, how is the host/manager to know what syntax to use for
placeholder tokens, when they are, in general, separately developed
codebases?

A related problem is how the host/manager knows whether and what
placeholder tokens are allowed in a given string? This question has been
moved out of the core library to be an industry-specific concern. Check
the docs in the MediaCreation repository; but in short, the
documentation of individual traits and properties dictate the allowed
tokens.

So we define an "OpenAssetIO format" for substitution tokens, which,
conveniently, is exactly the same as the libfmt format which,
conveniently again, almost exactly matches Python format string format.

However, we only support named (as opposed to positional) arguments,
and only officially support format specifiers for integer zero padding
(e.g. `{frame:04d}`).  Other format specifiers will work, but we're
reluctant to document support for them, since future substitution
implementations (e.g. in other languages) may not be able to use libfmt.

So add a `utils::substitute` function that takes a dictionary
mapping token names to values, and returns a substituted string, or an
error.

In particular, an error is thrown when insufficient variables are given
in the dictionary to fill all placeholders. This is undesirable, but
partial substitution is not supported in libfmt (or indeed, Python
format strings). Future work may add such support, but it is non-trivial
to add.

Signed-off-by: David Feltell <david.feltell@foundry.com>