[Lint] Disable internal BatchElementError deprecation warnings
Discovered whilst working on OpenAssetIO/usdOpenAssetIOResolver#39,
which required use of VFX Reference Platform CY23, i.e. GCC 11. We
currently work on CY22, i.e. GCC 9.3.

When bumping to GCC 11 `-Werror` causes the build to fail with
`-Werror=deprecated-declarations` because we deliberately include the
deprecated `BatchElementException` types in the binary, to satisfy
symbol discovery on Windows.

The fact this isn't flagged as an error when compiling with GCC 9.3 is
probably a deficiency in that compiler version.

Signed-off-by: David Feltell <david.feltell@foundry.com>