[Build] Improve speed of CMake configure
We were needlessly repeating the same work to compute platform warnings
and IPO capability in the `openassetio_set_default_target_properties`
function every time it was called.

On Linux the slowdown is barely noticeable. However, on Windows the
slowdown is frustrating. A re-configure of a hot build directory
goes down from 1 min 43 secs before this commit to a mere 8 secs, on
Windows.

Signed-off-by: David Feltell <david.feltell@foundry.com>