[CI] Allow use of Node 16
Our ABI diff check runs on our Docker image, which is based on the ASWF
VFX Reference Platform 2022 image, whose glibc version is rather old.
This makes it incompatible with Node 20. We have been holding back the
version of `actions/checkout` to ensure Node 16 was still used. However,
GitHub have started forcing Node 20 regardless. See
https://github.blog/changelog/2024-03-07-github-actions-all-actions-will-run-on-node20-instead-of-node16-by-default/

That link gives a way to buy more time on Node 16 via a
`ACTIONS_ALLOW_USE_UNSECURE_NODE_VERSION` env var, but it only seems to
be for self-hosted runners.

The real fix will be to upgrade our Docker image, which will be done in
#984.

Signed-off-by: David Feltell <david.feltell@foundry.com>