[Core] Add singular overload of managementPolicy
Part of #856. There are many cases where hosts only need to query a
single trait set.

So provide a convenient overload that takes a single trait set and
returns a single policy `TraitsData`.

Ensure the result from the manager plugin is bounds checked, to avoid
segfault. However, we don't use a custom human-readable exception, since
that case is truly exceptional - it is against the API contract and so
should never happen.

Signed-off-by: David Feltell <david.feltell@foundry.com>