[Build] Use shared libraries in Python wheel builds
Closes #1340. When OpenAssetIO is installed into a Python environment
using `pip install openassetio`, i.e. via our Python wheel, it would
install a single shared library - the Python extension module - with the
other OpenAssetIO libraries statically linked into it.

This meant there was nothing for C++ OpenAssetIO plugins to link
against, in terms of both library file, and exposed symbols. Given
differences in how libraries are loaded and symbols are exported, across
different platforms and different applications (e.g. Python vs. others),
we have to carefully structure our distribution.

Specifically, we must always build as shared libraries, so that C++
plugins can find their dependency, and we must bundle the libraries with
the Python distribution package.

On Windows, when building a default CMake install tree, we must
duplicate the shared libraries within the Python `site-packages`
directory, so that the dependencies can be found when Python is used to
load OpenAssetIO via the extension module. The CMake logic for doing
this duplication was already in place, but was a hacky solution to
Python linker errors from some time ago.

So consolidate the CMake logic for ensuring the `openassetio-core` and
`opeanssetio-python` targets are installed as part of the Python
distribution. Hence rename the `openassetio-python-module` install
`COMPONENT` to `openassetio-python-distribution` to better reflect its
purpose.

In `setup.py`, perform a full build of all targets, but only install the
`openassetio-python-distribution` component. This prevents irrelevant
artifacts, e.g. headers, from being bundled with the Python wheel.

Tweak the `setuptools` CMake preset used by `setup.py` so that
* On POSIX we don't bother with `.so.1.0.0`-style symlinks in the Python
  distribution.
* The libraries and binaries (binaries includes `.dll` files on
  Windows) are installed into a top-level `openassetio` directory in the
  install prefix. When building wheels, the install prefix reflects the
  updated Python `site-packages` directory. So the `openassetio`
  subdirectory is the distribution package.

Ensure a clear warning is displayed to anyone attempting to configure
OpenAssetIO to build as static libraries.

Clarify the CMake code comment for SimpleCppManager to reflect latest
investigations.

Signed-off-by: David Feltell <david.feltell@foundry.com>