[Build] Remove support for editable Python installs
Part of #1340. `pip install --editable` allows for a Python distribution
package to be installed as (essentially) a symlink to the sources it was
built from, allowing changes in the source tree to be reflected in the
Python environment, without requiring a reinstall. This can be very
useful for iterative development (e.g. TDD).

However, OpenAssetIO is now mostly in C++ with Python bindings, meaning
editable installs are not very useful, since most changes require a
rebuild/reinstall anyway.

Editable installs were only partially supported anyway, since .pyi stub
files are not handled properly.

The investigation in #1340 revealed that we will have to modify the
distribution to now bundle multiple shared libraries. This will
completely break editable installs, unless significant changes to
setup.py are made.

So given that editable installs are of little use, and are set to break,
disable support for them. Remove CI tests, and raise an exception if
attempted.

Support for editable installs will be further investigated in #1309.

Signed-off-by: David Feltell <david.feltell@foundry.com>