[Docs] Fix outdated BUILDING instructions
Discovered whilst working on #1340. Some of our build instructions are
out of date.

Requirements for CMake was too low, and pybind11-stubgen is a
build requirement by default.

The warnings relating to Python were out of date, since they were
written back when we used/recommended the `cpython` ConanCenter package.
This had a bunch of problems (see #977), so we moved to relying on the
system/user-provided Python.

Add some missing CMake options.

Remove docs for coverage and C API options. Coverage has never been
tested and should be considered unsupported. The C API is woefully
incomplete and needs rethink, so should be considered
deprecated/unsupported.

Signed-off-by: David Feltell <david.feltell@foundry.com>